# rgw: use chunked encoding to get partial results out faster