@mattbenjamin 2018-09-05
@robbat2 hey robbin, just to clarify
1. this works with the beast fe as well as civetweb?
2. do you speculate that different flushing behaviors will lead to shorter total response times?
Matt