@robbat2 2018-09-06
@mattbenjamin 
1. Beast definitely exhibits the same problem (proven w/ `ceph-nano`), but I didn't confirm that the fix solved it on Beast, as I did development on Luminous and forward-ported the fix, and couldn't spin up a test environment to the scale where the fix would show up in performance tests. The Chunked encoding itself should work fine as a couple of the Swift operations use it.
2. Yes, different flushing behavior will definitely shorten the total response time: by having less (but larger) packets or simply more packets in flight (if it doesn't wait for the TCP ACK before sending more).