@mattbenjamin 2018-11-08
http://pulpito.ceph.com/mbenjamin-2018-11-07_16:28:58-rgw-wip-chunked-encoding-distro-basic-smithi/